src/social/user/infra/api/signup/user_sign_up_request.py
from pydantic import BaseModel, Field


class UserSignupRequest(BaseModel):
    name: str = Field(examples=["John Doe"])
    username: str = Field(examples=["johndoe"])
    email: str = Field(examples=["johndoe@hotmail.com"])
