pyproject.toml
[project]
name = "social-network"
version = "0.5.1"
description = "Project of a social network to practice DDD"


dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.7",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "sqlalchemy>=2.0.36",
    "validators>=0.34.0",
]

requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
test = [
    "coverage>=7.6.9",
    "doublex>=1.9.6.1",
    "doublex-expects>=0.7.1",
    "expects>=0.9.0",
    "faker>=35.0.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-sugar>=1.0.0",
    "pytest-watch>=4.2.0",
    "pytest-xdist>=3.6.1",
    "testcontainers>=4.10.0",
]
lint = [
    "mypy>=1.13.0",
    "ruff>=0.8.3",
]
git = [
    "commitizen>=4.1.1",
    "cz-conventional-gitmoji>=0.6.1",
    "pre-commit>=4.1.0",
]
